Bug fix 1

+++ src/main/java/com/flickr4java/flickr/util/FileAuthStore.java
@@ -127,14 +127,15 @@ public class FileAuthStore implements AuthStore {
      * 
      * @see com.flickr4java.flickr.util.AuthStore#clearAll()
      */
-    // TODO: null check for dereference for auths local variable
     public void clearAll() {
         this.auths.clear();
         this.authsByUser.clear();
         File[] auths = this.authStoreDir.listFiles(new AuthFilenameFilter());
-        for (int i = 0; i < auths.length; i++) {
+        if (auths != null) {
+        	for (int i = 0; i < auths.length; i++) {
             auths[i].delete();
-        }
+        	}
+        }      
     }
 
     /*
